Refactoring to have it more visible to humans
