+++ app/src/main/java/cat/udl/hyperion/appmobils/kingdomcollector/other/MainActivity.java
@@ -24,7 +24,7 @@ import java.util.Objects;
 import cat.udl.hyperion.appmobils.kingdomcollector.R;
 import cat.udl.hyperion.appmobils.kingdomcollector.game.views.GameActivity;
 import cat.udl.hyperion.appmobils.kingdomcollector.other.auth.LastLoginCallback;
-import cat.udl.hyperion.appmobils.kingdomcollector.views.CardCollectionActivity;
+import cat.udl.hyperion.appmobils.kingdomcollector.collection.views.CardCollectionActivity;
 
 
 public class MainActivity extends AppCompatActivity {
